+++ spec/municipitaly2_spec.rb
@@ -2,6 +2,6 @@
 
 RSpec.describe Municipitaly2 do
   it 'has a version number' do
-    expect(Municipitaly2::VERSION).to be '2.0.0'
+    expect(Municipitaly2::VERSION).to be '2.0.3'
   end
 end