+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.3] - 2024-12-17
+
+- Search region by its name: `Municipitaly::Search.province_from_name('a province')`
+- Fixed and tested specs
+
 ## [2.0.2] - 2024-12-13
 
 - Added `csv` to spec as advised for ruby `3.4.0` onward